+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `whmcs-libvirt` will be documented in this file.
 
+## 0.1.0 - 2022-02-07
+
+- Display a table with all VM hosts and guests
+- Ability to Refresh Guests (retrieve all running guests from the active hosts)
+
 ## 0.0.1 - 2022-02-06
 
-- initial pre-release
+- Initial pre-release